requirements.txt
reportlab==4.0.4
pymongo==4.6.0
mysql-connector-python==8.2.0
Pillow==10.1.0
python-dateutil==2.8.2
pydantic==2.5.0
click==8.1.7
tkinter-tooltip==2.1.0